+++ src/main/java/blackjack/domain/card/CardDeck.java
@@ -1,7 +1,7 @@
 package blackjack.domain.card;
 
-import blackjack.domain.card.cardelement.Denomination;
-import blackjack.domain.card.cardelement.Suit;
+import blackjack.domain.card.element.Denomination;
+import blackjack.domain.card.element.Suit;
 import java.util.Collections;
 import java.util.Stack;
