+++ src/main/java/blackjack/domain/human/Human.java
@@ -2,8 +2,10 @@ package blackjack.domain.human;
 
 import static blackjack.util.Constants.BLACKJACK_NUMBER;
 
+import blackjack.domain.Point;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
+import blackjack.domain.human.element.Name;
 
 public abstract class Human {
     protected final Name name;
@@ -29,7 +31,7 @@ public abstract class Human {
     }
     
     public int getPoint() {
-        return cards.getPoint();
+        return Point.fromCards(cards).get();
     }
     
     public String getName() {
