+++ src/main/java/blackjack/domain/human/element/Name.java
@@ -1,4 +1,4 @@
-package blackjack.domain.human;
+package blackjack.domain.human.element;
 
 import java.util.regex.Pattern;
